+++ discovery-provider/flake8_plugins/flask_decorator_plugin/flask_decorator_plugin_test.py
@@ -1,6 +1,7 @@
 import ast
 
-from .flask_decorator_plugin import Plugin, Visitor
+from flask_decorator_plugin import Plugin
+from visitor import Visitor
 
 non_route_param_example = """
 @full_ns.route("/<string:track_id>/remixes")
